feat: dynamic navigation root
