+++ packages/navigation/src/index.tsx
@@ -1,3 +1,3 @@
 export { useWebviewNavigate } from './hooks/useWebviewNavigate';
 export { useCurrentUrl, LinkingConfig } from './hooks/useCurrentUrl';
-export * from './buildWebScreen';
+export { default as WebScreenNavigation } from './WebScreenNavigation';
